fix: fix some errors from merge
