+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "openbim-components",
-  "version": "1.4.0-alpha.7",
+  "version": "1.4.0-alpha.10",
   "main": "src/index.js",
   "author": "harry collin, antonio gonzalez viegas",
   "license": "MIT",
@@ -19,13 +19,12 @@
     "@tailwindcss/forms": "^0.5.3",
     "@types/earcut": "^2.1.1",
     "@types/jest": "27.0.0",
-    "@types/mapbox-gl": "^2.7.11",
     "@types/node": "^14.14.31",
     "@types/node-fetch": "^2.6.4",
     "@types/three": "0.160.0",
     "@typescript-eslint/eslint-plugin": "^4.27.0",
     "@typescript-eslint/parser": "^4.27.0",
-    "bim-fragment": "1.3.8",
+    "bim-fragment": "1.3.9",
     "cpy-cli": "^3.1.1",
     "eslint": "^7.28.0",
     "eslint-config-airbnb-base": "^14.2.1",
@@ -33,7 +32,6 @@
     "eslint-plugin-import": "^2.23.4",
     "eslint-plugin-prettier": "^3.4.0",
     "jest": "^27.0.4",
-    "mapbox-gl": "^2.15.0",
     "node-fetch": "^3.3.2",
     "prettier": "^2.3.1",
     "rollup": "^3.2.3",
@@ -59,8 +57,7 @@
     "unzipit": "^1.4.3"
   },
   "peerDependencies": {
-    "bim-fragment": "1.3.8",
-    "mapbox-gl": "^2.15.0",
+    "bim-fragment": "1.3.9",
     "three": "^0.160.1",
     "web-ifc": "^0.0.50"
   }
