+++ src/index.ts
@@ -7,7 +7,6 @@ export * from "./navigation";
 export * from "./annotation";
 export * from "./utils";
 export * from "./ifc";
-export * from "./integrations";
 export * from "./measurement";
 export * from "./import-export";
 export * from "./civil";
