--- src/integrations/index.ts
@@ -1,3 +1,0 @@
-export * from "./mapbox";
-
-// TODO: Remove this?
